+++ src/Engine/Graphics/Vis.cpp
@@ -72,9 +72,9 @@ Vis_ObjectInfo *Vis::DetermineFacetIntersection(BLVFace *face, Pid pid, float pi
             }
         }
     } else if (uCurrentlyLoadedLevelType == LEVEL_OUTDOOR) {
-        const std::vector<Vec3i> &v = pOutdoor->model(pid).pVertices;
+        const std::vector<Vec3f> &v = pOutdoor->model(pid).pVertices;
         for (unsigned i = 0; i < face->uNumVertices; ++i)
-            static_DetermineFacetIntersection_array_F8F200[i].vWorldPosition = v[face->pVertexIDs[i]].toFloat();
+            static_DetermineFacetIntersection_array_F8F200[i].vWorldPosition = v[face->pVertexIDs[i]];
     } else {
         assert(false);
     }
@@ -316,7 +316,7 @@ bool IsBModelVisible(BSPModel *model, int reachable_depth, bool *reachable) {
     float radius{ static_cast<float>(model->sBoundingRadius) };
     if (radius < 512.0f) radius = 512.0f;
 
-    return IsSphereInFrustum(model->vBoundingCenter.toFloat(), radius);
+    return IsSphereInFrustum(model->vBoundingCenter, radius);
 }
 
 bool IsSphereInFrustum(Vec3f center, float radius, Planef *frustum) {
